--- src/main/steps/index.utils.ts
@@ -1,5 +1,0 @@
-import { CaseWithId } from '../app/case/case';
-
-export const hasSubmittedAos = (userCase: CaseWithId): boolean => {
-  return Boolean(userCase.dateAosSubmitted);
-};
